starting to move envs into directories
